Refactor Lunabase communication and remove cakap2 protocol
- Replaced the cakap2 protocol with a new LunabaseConnection structure for handling communication.
- Simplified the PacketBuilder and LunabaseConn implementations.
- Removed the cakap2 crate and its associated files, including error handling and packet management.
- Updated Lunabase task to utilize the new LunabaseConnection for message handling and keep-alive functionality.
- Adjusted QUIC server listen method to accept u16 for port instead of u32.
- Modified tests to align with the new port type.

+++ misc/quic/src/lib.rs
@@ -342,7 +342,7 @@ impl<Outgoing: Encode + Decode<()>, Incoming: Encode + Decode<()>> QuicServer<Ou
 
     /// Listens on 0.0.0.0:{port}
     /// Accepts incoming connections
-    pub fn listen(port: u32, keep_alive_interval: Duration) -> Result<Self, Box<dyn std::error::Error + Send + Sync>> {
+    pub fn listen(port: u16, keep_alive_interval: Duration) -> Result<Self, Box<dyn std::error::Error + Send + Sync>> {
         Self::ensure_runtime();
         let (endpoint, _cert) = tasker::get_tokio_handle().block_on(async {
             make_server_endpoint(std::net::SocketAddr::V4(
